pyproject.toml
[project]
name = "mcp-server-kosa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain==0.3.13",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-openai==0.2.14",
    "langgraph==0.2.60",
    "mcp[cli]>=1.6.0",
    "docx2txt==0.8",
    "chroma-hnswlib==0.7.6",
    "chromadb==0.5.23",
    "langchain-chroma==0.1.4",
    "langchain-community==0.3.13",
    "langchain-core>=0.3.56",
    "numpy==1.26.4",
    "openai==1.58.1",
    "python-dotenv==1.0.1",
    "docx2txt==0.8",
    "onnxruntime==1.20.1",
    "ipykernel==6.29.5",
]
